fix error TURDEVSUP-366 + add blind offers in get active offers

+++ src/Domain/Entities/CampaignsManagement.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities
         public int Id { get; set; }
         public string? ExternalCampaignId { get; set; }
         public int? IdjobVacancy { get; set; }
-        public int? Status { get; set; }
+        public byte? Status { get; set; }
         public DateTime? LastModificationDate { get; set; }
         public int? ModificationReason { get; set; }
         public decimal? Budget { get; set; }
